Removed KuraRuntimeException from throws clauses
Checked exceptions must be included in the throws clause whereas Runtime
exceptions (aka Unchecked Exceptions) need not be included in the throws
clause as Unchecked Exceptions are mainly for programmatic errors. So,
this would be considered as redundant if included in the throws clause.
So, in case of this, it is better off to include Runtime exceptions in
javadoc comment, rather than in the throws clause.

Signed-off-by: Amit Kumar Mondal <admin@amitinside.com>