Minimized accessibility of classes and members
According to Effective Java by Joshua Bloch Item #13, the single most
important factor that distinguishes a well-designed module from a poorly
designed one is the degree to which the module hides its internal data
and other implementation details. If the classes are not a part of the
exported API, it is better off to provide package private (default)
access if and only if it is used within the classes of the same package.

Signed-off-by: Amit Kumar Mondal <admin@amitinside.com>