Refactoring the Apache Camel integration
This change upgrades to Apache Camel 2.17.2 and does remove the
unnecessary workarounds for older Camel versions. Languages and
components get auto-detected by OSGi now.

Also are the Kura related CamelRouter classes being split up in:
* A basic Kura based Camel component
* A basic XML routes based component
* A basic XML routes component which gets configured as OSGi service
component
* A generic Kura component based on a single XML snippet with UI support

This change also fixes the "sun.misc" import issue in a way that the
bootdelegation is not required.

In addition the following changes are done:
* Adds a bit of logging and fixed a possible NPE
* Splits of the GatewayRouter example to two: one Java, one XML approach
  
The KuraCloudComponent does now register with OSGi as soon as a
CloudService is being detected and will be available automatically
in the camel context.

The global cache, which leaked CloudClient instances, was removed and
integrated into the camel component, which will track clients and clean
up when it is stopped.

The deadlock while stopping the camel context, when the camel cloud
client was being used, is fixed by using a workaround which detects if
it the context is stopping.

Topics of the KuraCloudConsumer are being re-established when the
connection to the broker is being re-established.

Signed-off-by: Jens Reimann <jreimann@redhat.com>