#200: remove the need for bootdelegation to sun.misc
This change does remove the necessary to add the bootdelegation to
sun.misc by creating two fragments which do the proper wiring.

One fragment is extending the "system.bundle" to export "sun.misc" and
the other one is extending "camel-core" to import "sun.misc" as it
should be done by camel.

Still, this will only work with a JRE which actually provides "sun.misc"
and it will probably stop working with Java 9.

AFAIK the core issue will be fixed by Camel 2.18 in a future release.

Signed-off-by: Jens Reimann <jreimann@redhat.com>