#125: fix deadlock while shutting down Apache Camel router
This change fixes one part of the BundleContext issue.

While shutting down an Apache Camel router the shutdown runs
into a deadlock, which will be resolved by Camel after a
timeout by simply killing the thread.

Shutting down the camel context and its services is being
forked into another thread, while holding a mutex lock on the
camel context. At the same time the other thread shuts down all
camel components registered with the context and will also
shut down the camel based "cloud client". Which will then shut
down its routes, and with that make a call to the camel context
from this "shutdown thread", requiring a mutex lock on the same
context object. While the first thread it still waiting for the
second thread to finish.

It seems as there is no proper way to handle this issue. But is it
possible so try and detect if a shutdown is currently in progress. In
this case the camel cloud client shutdown now does not destroy its
routes, since they are stopped anyway by the whole shutdown process.

This prevents the shutdown thread trying to acquire the same sync lock.

Signed-off-by: Jens Reimann <jreimann@redhat.com>