Fix uncaught IllegalArgumentException
The kura.configuration property is expected to be a URI, during the
initial loading process it is converted from a String to a URI and then
passed into a File constructor.

This may result in an IllegalArgumentExecption when something is wrong
with the URI. However that exception is not caught.

This change cleans up the load process by opening the resource using
URL.openStream() and also catches any possible exception.

Signed-off-by: Jens Reimann <jreimann@redhat.com>