Change the build to a self contained manifest-first tycho build
This changes fixes point 1 in #265 by converting the bundles to a
manifest-first, self contained build based in Tycho.

In order to allow a self contained build a hosted P2 repository is
necessary. I am using a Package Drone hosted P2 repository for the
moment: [1]

This bug also depends on PR #428. As long as the Package Drone P2
repository is being used this issue does not block this change.

As long as PR #428 is merged, the URL can be changed to any other hosted
P2 repository.

[1] https://packagedrone.eclipse.org/channel/eclipse-kura-2.0/view

Signed-off-by: Jens Reimann <jreimann@redhat.com>