Fix a possible NPE and do minor cleanups (#415)
This change fixes a possible NPE which can be triggered if the web ui
is disabled. In that case the update method will try to access a field
which never got initialized. This is sovled by adding another check
if the web ui is enabled.

I addition the unused executor service was removed.

Signed-off-by: Jens Reimann <jreimann@redhat.com>