#189 implement CloudService with Camel backend
This change does implement the CloudService API and provides a cloud
service based on a camel router with XML route configuration. It uses
the SelfConfiguringComponent interface in order to create
configurations.

As of now two issues remain with Kura:
 * All components make the assumption that only one CloudService is
   present, this may cause a component to just get any random service
   assigned - #466
 * It is hard to configure the router XML in a single line text entry
   field - #345
   
At the moment the implementation does not use the CloudServiceFactory 
though.

Signed-off-by: Jens Reimann <jreimann@redhat.com>