#523 only calculate checksums once
This changed does restructure the build so that they get only calculated
once. Before that each device profile did generate the checksums for
all other device profiles in addition.

The main change is that the calls to "build_equinix_distrib.xml" get
shifted to the "package" phase and that there is only one call to the
checksum plugin in the "install" phase.

In order to not let the checksum plugin fail when do "deb" profile is
active, a dummy file is being generated early in the build.

Signed-off-by: Jens Reimann <jreimann@redhat.com>