Switch to using "?" as location for configurations
This commit changes all locations where Kura gets configurations from
the ConfigurationAdmin to use the location "?" which disables the
location binding. As @amitjoy pointed out this change is preferred over
the older version of passing in `null`.

In Kura one service will create configurations for services which reside
in other bundles, so access must be possible for them.

Signed-off-by: Jens Reimann <jreimann@redhat.com>