Start the Web UI component immediately
The Web UI component should be started immediately on startup. If the
immediate attribute is not set, the component will become a "delayed"
component and only be activated once there is a reference to it (also
see 112.5.4 of the OSGi compendium specification). However no one will
actively reference this component.

Signed-off-by: Jens Reimann <jreimann@redhat.com>