Enable BREE validation on Travis CI build
This change adds a BREE validation profile to the existing tycho build.
A profile can be used to enable this validation and by default this
profile is not active.

This profile is turned on during the Travis CI build, so that code can
be validated against the actual configured BREE.

Signed-off-by: Jens Reimann <jreimann@redhat.com>