#641: Check connection state before connecting
This change does perform a check first if the connection is establed.

This is not a 100% guarantee that the next call will succeed, bit it
should indeed catch most of the cases. If, on the other hand the state
is false and suddenly pops up, the callback should cover that case.

Signed-off-by: Jens Reimann <jreimann@redhat.com>