Fix issue #691
This change does fix issue #691 by re-implementing the way the UI
backend does locate the components the communication stack is made up.

Instead of doing another round-trip to the server the backend extracts
the factory id from the service, locates the factory instance and
fills the result in one step.

This change also reduces a few assumptions made on the services, like
the requirement on a CloudService that is has to be configurable, the
requirement on the factory to be implemented using OSGi DS.

Signed-off-by: Jens Reimann <jreimann@redhat.com>