Allow the Web UI to configure un-registered CloudService instances
This change does fix an issue with the new multi-cloud-service feature
which might occur when services are no longer registered with OSGi.

The problem is that the Web UI does require a CloudService to be
registered in order to detect it. But if there is a configuration issue
with the cloud service, then it doesn't get registered and the Web UI
will not pick it up. So a configuration which causes a service to
unregister cannot be fixed again.

This change now introduces a new method in the CloudServiceFactory which
must report back all configured CloudService instances according to
this cloud service factory instance. It is the responsiblity of the
factory to enumerate its, and only its, managed CloudService instances.

The Web UI will iterate over all CloudServiceFactory instances and
gather a list of all CloudService instances.

This way it is possible to re-use the factory ID, assigned with that
CloudService and each cloud service factory can actually probe the
ConfigurationService for configured (not registered) CloudService
instances.

A new column was added to show if the service is actually registered.

Signed-off-by: Jens Reimann <jreimann@redhat.com>