execute javadoc generation in the "site" phase
This change is necessary due to the fact that the current javadoc
configuration is run in the aggregator project and is configured to
fork the generate-source phase, which then triggers the build, which may
run into the issue that depencies are missing, since they haven't been
compiled yet.

The simple solution is to follow the standard maven javadoc generation
and keep the javadoc generation with the "site" phase, so that the
generation will take place **after** the main build is done.

This requires the maven command to the called with "site" in addition
if the javadoc should be generated.

Signed-off-by: Jens Reimann <jreimann@redhat.com>