Add support for forwarding OSGI events to the Web Ui
A subset of the OSGI EventAdmin events is now forwarded to the web ui using long polling.
The subset of events forwarded is defined by the ForwardedEventTopic enum.
The cloud services, bundles, packages and status sections are now refreshed on event.
The icon representing the cloud connection status in the left panel is now updated on event.

Signed-off-by: nicolatimeus <nicola.timeus@eurotech.com>