Work around Kura sending empty string instead of null for Camel deps
Once the service has been configured there is now way
with Kura to reset the value (to unset it) although it is optional.
In this case the value will be reported as "" (empty string)
instead of `null`.

This causes the Camel component to look for a component with an empty
name, which never exists.

Signed-off-by: Jens Reimann <jreimann@redhat.com>