Changes to prevent exception of wires panel refresh.
It can happen that, with heavy components, the getWiresConfigurationInternal() invocation
at the end of updateWireConfiguration() determines an internal error, probably due to a refresh
when the component is not effectively up and running.
The changes try to prevent this refresh, as not really needed.
Small change in WireComponentTrackerCustomizer.java to try to prevent a ConcurrentModificationException.

Signed-off-by: MMaiero <matteo.maiero@eurotech.com>