Revert working directory ignored in CMD-V1
This fixes #1558 by partially reverting commit 1c577d8.
With that commit the working directory specified in the CMD-V1 EXEC request is ignored.
It was introdouced as a workaround to support the Windows port where the cloud platform
always send /tmp as the working directory.
As there's no Windows port and the right thing to do is fixing the cloud platform
the old behavior is restored.

Signed-off-by: Cristiano De Alti <cristiano.dealti@eurotech.com>