Fix PDP auth with CID 1 in 4G. This is an attempt to fix #2264.
It is not perfect and relies on the modem monitor to recover the modem
if something goes wrong.
It applies to Telit products LE910-EU1 and LE910-V2 EU/NA.
PDP context 1 is rarely used and there should not be any
performance impact for the other contexts.
See also comment in TelitLe910v2.java.
* AT+CGDCONT removed from chat script
* PDP context and auth set only if config changes
* Improved tip in UI for LTE modems
Configuring context 1 it's likely to require a modem reset before
the changes take effect.
NOTE: changes to PDP auth type and credentials only take effect
if the APN is changed too.

Signed-off-by: Cristiano De Alti <cristiano.dealti@eurotech.com>