Modified net configuration read to be more resilient in case of interface disalignmemnt.
When the status of an interface is parsed, but this is not tracked by network.interfaces file, its status is
defaulted to static. This is not the safest approach because that means that an ip and subnet are requested.

The PR changes the default when parsing the network interfaces: the interfaces that are not tracked are
managed as in dhcp.

Signed-off-by: Maiero <matteo.maiero@eurotech.com>