Prevent unnecessary modifications inside the packages folder.
Prevent changes to the dpa.properties file if properties do not change.
Save .dp files with consistent naming "NAME_VERSION.dp".
Fixed a bug in which installed .dp files would be copied to a new dp
file named "NAME_VERSION.dp" on startup if they are not already named
this way.

Signed-off-by: Andrej Krota <andrej.krota@gmail.com>